RM #17952 AVAILABITY REQUEST : error message if availableStock < qty - reservedQty on saleOrderLine
