+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased 5.1.7]
 ## Features
-- AVAILABITY REQUEST : error message if availableStock is less than realQty
+- AVAILABITY REQUEST : error message if availableStock is less than qty - reservedQty on saleOrderLine
 - USER : filter assigned to user in partner form
 - MRP FORECAST :change a status of MrpForecast object from 'Draft' to 'Confirmed' in demo data.
 - TRACKING : Added all fields for all apps and also for company's configuration
