+++ axelor-stock/src/main/java/com/axelor/apps/stock/exception/IExceptionMessage.java
@@ -185,7 +185,4 @@ public interface IExceptionMessage {
   String TRACK_NUMBER_WIZARD_TITLE = /*$$(*/ "Enter tracking numbers" /*)*/;
   String TRACK_NUMBER_WIZARD_NO_RECORD_ADDED_ERROR = /*$$(*/ "No Tracking Numbers Added" /*)*/;
   String TRACK_NUMBER_DATE_MISSING = /*$$(*/ "Please filled estimated delivery date" /*)*/;
-
-  static final String STOCK_MOVE_VERIFY_PRODUCT_STOCK_ERROR = /*$$(*/
-      "Product stock for %s is not enough for availability request" /*)*/;
 }
