+++ axelor-stock/src/main/resources/domains/AppStock.xml
@@ -10,14 +10,13 @@
 		<boolean name="isSeparateShipmentOperations" title="Separate the shipment operations"/>
 		<boolean name="isRequiredShipmentSupplierDetails" title="Require to fill the shipment informations"/>
         <many-to-one name="economicArea" ref="com.axelor.apps.base.db.EconomicArea"/>
-        <boolean name="isVerifyProductStock" title="Verify product stock before availabity request"/>
+
         <track>
         	<field name="isEnabledProductDescriptionCopy" on="UPDATE"/>
         	<field name="requireToFillConformityOnReceipt" on="UPDATE"/>
         	<field name="isSeparateShipmentOperations" on="UPDATE"/>
         	<field name="isRequiredShipmentSupplierDetails" on="UPDATE"/>
         	<field name="economicArea" on="UPDATE"/>
-        	<field name="isVerifyProductStock" on="UPDATE"/>
         </track>
 	</entity>
