+++ axelor-stock/src/main/resources/i18n/messages_en.csv
@@ -55,6 +55,7 @@
 "Both source location address and delivery (to) address are required to define the delivery itinerary",,,
 "Buyer",,,
 "Buyer/Seller",,,
+"CSV",,,
 "Can't delete a planned stock move",,,
 "Can't realize this stock move because of the ongoing inventory %s.",,,
 "Cancel",,,
@@ -389,6 +390,7 @@
 "Outgoing move information",,,
 "Outgoing move lines",,,
 "Own propulsion",,,
+"PDF",,,
 "Pack price select",,,
 "Packing line",,,
 "Packing lines",,,
@@ -468,7 +470,6 @@
 "Product model",,,
 "Product name",,,
 "Product quantity in other stock locations",,,
-"Product stock for %s is not enough for availability request",,,
 "Product tracking number",,,
 "Product type",,,
 "Product's stocks %s (%s) are not in sufficient quantity to realize the delivery",,,
@@ -740,13 +741,13 @@
 "Validated",,,
 "Validated By",,,
 "Validated On",,,
-"Verify product stock before availabity request",,,
 "View Direction",,,
 "Virtual",,,
 "Virtual stock locations informations",,,
 "Warranty expiration date",,,
 "Wording",,,
 "Workshop",,,
+"XLS",,,
 "Yes",,,
 "You must configure a Stock module for the company %s",,,
 "You must configure a customer virtual stock location for the company %s",,,
