+++ axelor-stock/src/main/resources/i18n/messages_fr.csv
@@ -55,6 +55,7 @@
 "Both source location address and delivery (to) address are required to define the delivery itinerary","Les champs Emplacement source et A l'adresse sont requis pour définir l'itinéraire de livraison.",,
 "Buyer","Acheteur",,
 "Buyer/Seller","Acheteur/Vendeur",,
+"CSV",,,
 "Can't delete a planned stock move","Il n'est pas possible de supprimer un mouvement de stock planifié",,
 "Can't realize this stock move because of the ongoing inventory %s.","Impossible de réaliser ce mouvement de stock à cause de l’inventaire en cours %s.",,
 "Cancel","Annuler",,
@@ -389,6 +390,7 @@
 "Outgoing move information","Informations BL",,
 "Outgoing move lines","Lignes de BL",,
 "Own propulsion","Propulsion propre",,
+"PDF",,,
 "Pack price select",,,
 "Packing line","Ligne de colisage",,
 "Packing lines","Lignes de colisage",,
@@ -468,7 +470,6 @@
 "Product model","Modèle de produit",,
 "Product name","Nom du produit",,
 "Product quantity in other stock locations","Quantité du produit dans les autres emplacements",,
-"Product stock for %s is not enough for availability request","Le stock est insuffisant pour le(s) produit(s) %s pour une demande de mise à disposition",,
 "Product tracking number","No de suivi du produit",,
 "Product type","Type de produit",,
 "Product's stocks %s (%s) are not in sufficient quantity to realize the delivery","Les stocks du produit %s (%s) sont insuffisants pour réaliser la livraison",,
@@ -740,13 +741,13 @@
 "Validated","Validé",,
 "Validated By","Validé par",,
 "Validated On","Validé le",,
-"Verify product stock before availabity request","Vérifier les stocks avant une demande de mise à dispo",,
 "View Direction","Voir itinéraire",,
 "Virtual","Virtuel",,
 "Virtual stock locations informations","Informations Stocks Virtuels",,
 "Warranty expiration date","Date de fin de garantie",,
 "Wording","Libellé",,
 "Workshop",,,
+"XLS",,,
 "Yes","Oui",,
 "You must configure a Stock module for the company %s","Veuillez configurer le module stock pour la société %s",,
 "You must configure a customer virtual stock location for the company %s","Veuillez configurer un Emplacement Virtuel Client pour la société %s",,
