+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/exception/IExceptionMessage.java
@@ -214,4 +214,7 @@ public interface IExceptionMessage {
 
   public static final String SALE_ORDER_STOCK_MOVE_CREATED = /*$$(*/
       "Stock move %s has been created for this sale order" /*)*/;
+
+  static final String STOCK_MOVE_VERIFY_PRODUCT_STOCK_ERROR = /*$$(*/
+      "Product stock for %s is not enough for availability request" /*)*/;
 }
