+++ axelor-supplychain/src/main/resources/domains/AppSupplychain.xml
@@ -43,6 +43,7 @@
         <boolean name="intercoPurchaseOrderCreateRequested" title="Create interco purchase quotation as requested"/>
         <boolean name="intercoInvoiceCreateValidated" title="Create interco invoice as validated"/>
     <boolean name="blockDeallocationOnAvailabilityRequest" title="Block deallocation on availability request"/>
+    	<boolean name="isVerifyProductStock" title="Verify product stock before availabity request"/>
     
     <track>
     	<field name="manageInvoicedAmountByLine" on="UPDATE"/>
@@ -79,6 +80,7 @@
     	<field name="intercoPurchaseOrderCreateRequested" on="UPDATE"/>
     	<field name="intercoInvoiceCreateValidated" on="UPDATE"/>
     	<field name="blockDeallocationOnAvailabilityRequest" on="UPDATE"/>
+    	<field name="isVerifyProductStock" on="UPDATE"/>
     </track>
 	</entity>
