+++ axelor-supplychain/src/main/resources/i18n/messages.csv
@@ -320,6 +320,7 @@
 "Product is missing.",,,
 "Product launches/withdrawal each month",,,
 "Product name",,,
+"Product stock for %s is not enough for availability request",,,
 "Product type",,,
 "Product type distribution",,,
 "Products",,,
@@ -445,6 +446,7 @@
 "Usable on purchase order",,,
 "Usable on sale order",,,
 "Used Credit",,,
+"Verify product stock before availabity request",,,
 "You have to choose at least one incoming stock move",,,
 "You have to choose at least one outgoing stock move",,,
 "You must configure a Supply chain module for the company %s",,,
