+++ axelor-supplychain/src/main/resources/i18n/messages_fr.csv
@@ -320,6 +320,7 @@
 "Product is missing.","Il manque le produit.",,
 "Product launches/withdrawal each month","Lancement/Retrait produits par mois",,
 "Product name","Nom du produit",,
+"Product stock for %s is not enough for availability request","Le stock est insuffisant pour le(s) produit(s) %s pour une demande de mise à disposition",,
 "Product type",,,
 "Product type distribution","Distribution par type de produit",,
 "Products","Produits",,
@@ -445,6 +446,7 @@
 "Usable on purchase order","Utilisable pour les achats",,
 "Usable on sale order","Utilisable pour les ventes",,
 "Used Credit","Encours utilisé",,
+"Verify product stock before availabity request",,,
 "You have to choose at least one incoming stock move","Vous devez choisir au moins un BR",,
 "You have to choose at least one outgoing stock move","Vous devez choisir au moins un BL",,
 "You must configure a Supply chain module for the company %s",,,
