+++ axelor-supplychain/src/main/resources/views/AppSupplychain.xml
@@ -28,6 +28,7 @@
 		<panel name="stackPanel" title="Stock">
 			<field name="generateInvoiceFromStockMove" widget="boolean-switch" colSpan="4"/>
 			<field name="manageStockReservation" widget="boolean-switch" colSpan="4"/>
+			<field name="isVerifyProductStock" widget="boolean-switch" colSpan="4"/>
 		</panel>
 		<panel name="intercoPanel" title="Interco">
 			<field name="intercoFromPurchase" widget="boolean-switch"/>
